+++ src/renderer/components/screens/MainScreen.tsx
@@ -42,8 +42,6 @@ import { ScreenContext } from '../../contexts/ScreenContext'
 
 import type { T } from '@deltachat/jsonrpc-client'
 
-export type AlternativeView = 'global-gallery' | null
-
 type Props = {
   accountId?: number
 }
@@ -57,7 +55,14 @@ export default function MainScreen({ accountId }: Props) {
   const [queryStr, setQueryStr] = useState('')
   const [queryChatId, setQueryChatId] = useState<null | number>(null)
   const [archivedChatsSelected, setArchivedChatsSelected] = useState(false)
-  const { activeView, chatId, chat, selectChat, unselectChat } = useChat()
+  const {
+    activeView,
+    chatId,
+    chat,
+    alternativeView,
+    selectChat,
+    unselectChat,
+  } = useChat()
   const { smallScreenMode } = useContext(ScreenContext)
 
   // Small hack/misuse of keyBindingAction to setArchivedChatsSelected from
@@ -69,21 +74,6 @@ export default function MainScreen({ accountId }: Props) {
     setArchivedChatsSelected(false)
   )
 
-  const [alternativeView, setAlternativeView] = useState<AlternativeView>(null)
-  useEffect(() => {
-    if (chatId) {
-      setAlternativeView(null)
-    }
-  }, [chatId])
-  useKeyBindingAction(KeybindAction.GlobalGallery_Open, () => {
-    unselectChat()
-    setAlternativeView('global-gallery')
-  })
-  useKeyBindingAction(KeybindAction.GlobalMap_Open, () => {
-    unselectChat()
-    openMapWebxdc()
-  })
-
   const chatListShouldBeHidden =
     smallScreenMode && (chatId !== undefined || alternativeView !== null)
   const messageSectionShouldBeHidden =
@@ -91,7 +81,6 @@ export default function MainScreen({ accountId }: Props) {
 
   const onBackButton = () => {
     unselectChat()
-    setAlternativeView(null)
   }
 
   const onChatClick = (chatId: number) => {
@@ -417,7 +406,7 @@ function ChatNavButtons() {
         </Button>
         {settingsStore?.desktopSettings.enableOnDemandLocationStreaming && (
           <Button
-            onClick={() => openMapWebxdc(chatId)}
+            onClick={() => openMapWebxdc(selectedAccountId(), chatId)}
             active={activeView === ChatView.Map}
             aria-label={tx('tab_map')}
             className='navbar-button'