+++ src/renderer/hooks/chat/useSelectLastChat.ts
@@ -1,17 +1,22 @@
-import { useCallback, useEffect } from 'react'
+import { useCallback, useContext, useEffect } from 'react'
 
 import useChat from './useChat'
 import useHasChanged from '../useHasChanged'
 import { BackendRemote } from '../../backend-com'
 import { getLastChatId } from '../../backend/chat'
+import { ScreenContext } from '../../contexts/ScreenContext'
 
 /**
  * Detects account changes and selects last known, selected chat for the user
  * for this account.
  */
 export default function useSelectLastChat(accountId?: number) {
+  const { smallScreenMode } = useContext(ScreenContext)
   const hasAccountIdChanged = useHasChanged(accountId)
-  const { selectChat } = useChat()
+  const hasSmallScreenModeChanged = useHasChanged(smallScreenMode)
+  const { selectChat, chatId, alternativeView } = useChat()
+
+  const smallScreenModeChatListVisible = !(chatId || alternativeView)
 
   const selectLastChat = useCallback(async () => {
     if (!accountId) {
@@ -22,15 +27,26 @@ export default function useSelectLastChat(accountId?: number) {
 
     if (account.kind === 'Configured') {
       const lastChatId = await getLastChatId(accountId)
-      if (lastChatId) {
+      if (lastChatId && !smallScreenMode) {
         await selectChat(accountId, lastChatId)
       }
     }
-  }, [accountId, selectChat])
+  }, [accountId, selectChat, smallScreenMode])
 
   useEffect(() => {
-    if (hasAccountIdChanged) {
+    if (
+      hasAccountIdChanged ||
+      (hasSmallScreenModeChanged &&
+        !smallScreenMode &&
+        smallScreenModeChatListVisible) /* if small screen mode changed to false and it was not inside of a chat or global gallery */
+    ) {
       selectLastChat()
     }
-  }, [hasAccountIdChanged, selectLastChat])
+  }, [
+    hasAccountIdChanged,
+    hasSmallScreenModeChanged,
+    smallScreenModeChatListVisible,
+    smallScreenMode,
+    selectLastChat,
+  ])
 }